refactor: improve readability and maintainability of dataset handling
